769e70

+++ app/models/event.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 
 from .base import Base
 
@@ -11,6 +11,6 @@ class Event(Base):
     __tablename__ = "events"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     event = Column(String, nullable=False)
     ts = Column(DateTime, default=lambda: datetime.now(timezone.utc))
