+++ app/models/partner_order.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 from datetime import datetime, timezone
 
 from app.models.base import Base
@@ -8,7 +8,7 @@ class PartnerOrder(Base):
     __tablename__ = "partner_orders"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     order_id = Column(String, unique=True)
     protocol_id = Column(Integer)
     price_kopeks = Column(Integer)
