+++ app/models/photo_usage.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timezone
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 
 from .base import Base
 
@@ -9,7 +9,7 @@ class PhotoUsage(Base):
 
     __tablename__ = "photo_usage"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     month = Column(String(7), primary_key=True)
     used = Column(Integer, nullable=False, server_default="0")
     updated_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
